# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/v1/dns/location": {
            "post": {
                "description": "Calculate the location of the databank to upload gathered data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "DNS"
                ],
                "summary": "Calculate Location",
                "operationId": "dns",
                "parameters": [
                    {
                        "description": "DNS",
                        "name": "dns",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.DNSRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.DNSResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/routes.HTTPError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.DNSRequest": {
            "type": "object",
            "properties": {
                "vel": {
                    "type": "string"
                },
                "x": {
                    "type": "string"
                },
                "y": {
                    "type": "string"
                },
                "z": {
                    "type": "string"
                }
            }
        },
        "models.DNSResponse": {
            "type": "object",
            "properties": {
                "loc": {
                    "type": "string"
                }
            }
        },
        "routes.HTTPError": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        }
    }
}